Reorganize sdec into more convential python module
