+++ sdec/controller.py
@@ -11,7 +11,7 @@
 ####################################################################################
 # Imports                                                                          #
 ####################################################################################
-import sensor_conv
+from sdec import sensor_conv
 
 
 ####################################################################################
