+++ sdec/valveController.py
@@ -18,8 +18,8 @@ import serial.tools.list_ports
 #################################################################################### 
 # Project Modules                                                                  #
 #################################################################################### 
-import commands
-from   config   import *
+from sdec import commands
+from   sdec.config   import *
 
 
 ####################################################################################
